git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13283 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/mainwindow.cpp
@@ -225,8 +225,8 @@ void MainWindow::save(QString filename)
 
 void MainWindow::updateAttributes()
 {
-  m_ui.attributes->setRowCount(m_graph->getSizeAttributes());
-  for (size_t i = 0; i < m_graph->getSizeAttributes(); ++i)
+  m_ui.attributes->setRowCount((int)m_graph->getSizeAttributes());
+  for (int i = 0; i < m_graph->getSizeAttributes(); ++i)
   {
     for (int j = 0; j < m_ui.attributes->columnCount(); ++j)
     {
@@ -273,8 +273,8 @@ void MainWindow::updateValues()
     m_graph->calcAttrDistr(attributes[0], valueDistribution);
 
     Attribute *attribute = m_graph->getAttribute(attributes[0]);
-    m_ui.domain->setRowCount(attribute->getSizeCurValues());
-    for (size_t i = 0; i < attribute->getSizeCurValues(); ++i)
+    m_ui.domain->setRowCount((int)attribute->getSizeCurValues());
+    for (int i = 0; i < attribute->getSizeCurValues(); ++i)
     {
       Value *value = attribute->getCurValue(i);
